Move protobuf content to shared package.
