+++ core/shared/protobuf.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package connector
+package shared
 
 import (
 	"github.com/kiegroup/kogito-operator/api"
